+++ multidownloader/evm_multidownloader_test.go
@@ -317,7 +317,6 @@ func TestEVMMultidownloader_Initialize(t *testing.T) {
 		require.Error(t, err)
 		require.Contains(t, err.Error(), "already initialized")
 	})
-
 }
 
 /*
